refactor: base_font_size 从 uint16 改为 float，支持浮点精度
- XMConfig.base_font_size 类型改为 float
- StyleResolver 构造函数和成员变量适配 float
- normalize_font_size 保留浮点精度输出（最多 2 位小数，去除尾部零）
- 新增 CSSFontSizeFloatBase、CSSFontSizeFloatPt 测试
- 87 个测试全部通过

+++ core/include/xmarkup/xmarkup.h
@@ -105,7 +105,7 @@ typedef struct XMResult {
 typedef struct XMConfig {
     uint8_t  enable_autocorrect;
     uint16_t max_nesting_depth;
-    uint16_t base_font_size;
+    float    base_font_size;       /* 基准字号（px），支持浮点精度 */
 } XMConfig;
 
 /* 不透明解析器句柄 */
